requirements.txt
python-dotenv==0.17.1
PySide2==5.15.2